Restructured tests, they should work now
