--- ask/tests/test_ask.py
@@ -1,11 +1,0 @@
-import ask.utilities.tests as utilities_tests
-import ask.transpiler.tests as transpiler_test
-
-
-def main():
-	utilities_tests.main()
-	transpiler_test.main()
-
-
-if __name__ == '__main__':
-	main()
