+++ ask/tests/transpiler/test_lexer.py
@@ -1,47 +1,9 @@
 import unittest
 
-from ask.transpiler.utilities import lexer_utils
+from ask.transpiler import lexer
 
 
-class TestAskTranspilerUtilitiesLexerUtils(unittest.TestCase):
-	# tokens_grouped_by_lines()
-	def test_tokens_grouped_by_lines(self):
-		tokens = [
-			['VAR', 'variable'],
-			['FORMAT', '\n'],
-			['VAR', 'variable'],
-			['FORMAT', '\n'],
-			['FORMAT', '\t'],
-			['VAR', 'variable'],
-			['FORMAT', '\n'],
-			['FORMAT', '\t'],
-			['FORMAT', '\t'],
-			['VAR', 'variable'],
-		]
-
-		expected = [
-			[
-				['VAR', 'variable'],
-			],
-			[
-				['FORMAT', '\n'],
-				['VAR', 'variable'],
-			],
-			[
-				['FORMAT', '\n'],
-				['FORMAT', '\t'],
-				['VAR', 'variable'],
-			],
-			[
-				['FORMAT', '\n'],
-				['FORMAT', '\t'],
-				['FORMAT', '\t'],
-				['VAR', 'variable'],
-			],
-		]
-
-		self.assertEqual(expected, lexer_utils.tokens_grouped_by_lines(tokens))
-
+class TestAskTranspilerLexer(unittest.TestCase):
 	# insert_indentation_group_markers()
 	def test_insert_indention_group_markers(self):
 		tokens = [
@@ -100,12 +62,8 @@ class TestAskTranspilerUtilitiesLexerUtils(unittest.TestCase):
 			['GROUP', 'end']
 		]
 
-		self.assertEqual(expected, lexer_utils.insert_indention_group_markers(tokens))
-
-
-def main():
-	unittest.main()
+		self.assertEqual(expected, lexer.insert_indention_group_markers(tokens))
 
 
 if __name__ == '__main__':
-	main()
+	unittest.main()
