+++ ask/tests/transpiler/test_parser.py
@@ -4,11 +4,8 @@ from ask.transpiler import parser
 
 
 class TestAskTranspilerParser(unittest.TestCase):
-
-
-def main():
-	unittest.main()
+	pass
 
 
 if __name__ == '__main__':
-	main()
+	unittest.main()
