+++ ask/tests/transpiler/test_transpiler.py
@@ -4,11 +4,8 @@ from ask.transpiler import transpiler
 
 
 class TestAskTranspilerTranspiler(unittest.TestCase):
-
-
-def main():
-	unittest.main()
+	pass
 
 
 if __name__ == '__main__':
-	main()
+	unittest.main()
