+++ ask/tests/transpiler/utilities/test_lexer_utils.py
@@ -1,0 +1,47 @@
+import unittest
+
+from ask.transpiler.utilities import lexer_utils
+
+
+class TestAskTranspilerUtilitiesLexerUtils(unittest.TestCase):
+	# tokens_grouped_by_lines()
+	def test_tokens_grouped_by_lines(self):
+		tokens = [
+			['VAR', 'variable'],
+			['FORMAT', '\n'],
+			['VAR', 'variable'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['VAR', 'variable'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FORMAT', '\t'],
+			['VAR', 'variable'],
+		]
+
+		expected = [
+			[
+				['VAR', 'variable'],
+			],
+			[
+				['FORMAT', '\n'],
+				['VAR', 'variable'],
+			],
+			[
+				['FORMAT', '\n'],
+				['FORMAT', '\t'],
+				['VAR', 'variable'],
+			],
+			[
+				['FORMAT', '\n'],
+				['FORMAT', '\t'],
+				['FORMAT', '\t'],
+				['VAR', 'variable'],
+			],
+		]
+
+		self.assertEqual(expected, lexer_utils.tokens_grouped_by_lines(tokens))
+
+
+if __name__ == '__main__':
+	unittest.main()
