+++ ask/tests/transpiler/utilities/test_small_transpilers.py
@@ -4,11 +4,8 @@ from ask.transpiler.utilities import small_transpilers
 
 
 class TestAskTranspilerUtilitiesSmallTranspilers(unittest.TestCase):
-
-
-def main():
-	unittest.main()
+	pass
 
 
 if __name__ == '__main__':
-	main()
+	unittest.main()
