+++ ask/tests/utilities/test_utils.py
@@ -4,11 +4,8 @@ from ask.utilities import utils
 
 
 class TestAskUtilitiesUtils(unittest.TestCase):
-
-
-def main():
-	unittest.main()
+	pass
 
 
 if __name__ == '__main__':
-	main()
+	unittest.main()
