--- ask/transpiler/tests/__init__.py
@@ -1,9 +1,0 @@
-import ask.transpiler.utilities.tests as transpiler_utilities_tests
-
-
-def main():
-	transpiler_utilities_tests.main()
-
-
-if __name__ == '__main__':
-	main()
