--- ask/transpiler/tests/test_lexer.py
@@ -1,14 +1,0 @@
-import unittest
-
-from ask.transpiler import lexer
-
-
-class TestAskTranspilerLexer(unittest.TestCase):
-
-
-def main():
-	unittest.main()
-
-
-if __name__ == '__main__':
-	main()
