--- ask/transpiler/utilities/tests/__init__.py
@@ -1,10 +1,0 @@
-from ask.transpiler.utilities.tests import test_lexer_utils, test_parser_utils
-
-
-def main():
-	test_lexer_utils.main()
-	test_parser_utils.main()
-
-
-if __name__ == '__main__':
-	main()
