--- ask/utilities/tests/__init__.py
@@ -1,9 +1,0 @@
-from ask.utilities.tests import test_file_utils
-
-
-def main():
-	test_file_utils.main()
-
-
-if __name__ == '__main__':
-	main()
